Fixed and changed a whole bunch of things regarding combat
Added trails, changed the behavior of melee weapon, changed behavior of shielders and chargers. Added lights to enemies, added emission to shielder, added sounds to ammo pickups, fixed player being able to hit ammo, fixed chargers killing other AI enemies, added a pause when reading logbooks. Made changes to audio, volume and pitch levels

+++ G.A.R.A/Assets/Scripts/AI/AIDirector/PlayerReader.cs
@@ -36,7 +36,7 @@ public class PlayerReader : MonoBehaviour
         
     }
 
-    public int GetPlayerHP()
+    public float GetPlayerHP()
     {
         return playerStats.health;
     }
